+++ src/components/Item/__tests__/Discount.test.js
