+++ src/components/Item/__tests__/SmallText.test.js
