+++ src/components/Item/__tests__/Thumbnail.test.js
