+++ src/components/ItemsList/ItemsList.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import Item from '../Item2';
+import Item from '../Item';
 import CartItemControls from '../CartItemControls';
 import Text from '../Text';
 import List from './List';
